+++ rlclientlib/rlclientlib.vcxproj.filters
@@ -46,6 +46,7 @@
     <ClCompile Include="slates_response.cc" />
     <ClCompile Include="logger\logger_facade.cc" />
     <ClCompile Include="serialization\payload_serializer.cc" />
+    <ClCompile Include="generic_event.cc" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\include\api_status.h" />
@@ -114,6 +115,7 @@
     <ClInclude Include="..\include\errors_data.h" />
     <ClInclude Include="logger\logger_facade.h" />
     <ClInclude Include="serialization\payload_serializer.h" />
+    <ClInclude Include="generic_event.h" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
